Move StateList to __init__ and flake8 code cleanup
