+++ insteonplm/devices/unknowndevice.py
@@ -3,7 +3,8 @@ from insteonplm.devices import DeviceBase
 
 
 class UnknownDevice(DeviceBase):
-    """
+    """Unknown Device Class.
+
     Unknown Device used when only the device address is known but no other
     information. Available methods:
         id_request()
@@ -20,6 +21,7 @@ class UnknownDevice(DeviceBase):
     """
     def __init__(self, plm, address, cat=None, subcat=None, product_key=0x00,
                  description='', model=''):
+        """Initalize the UnknownDevice Class."""
         self._noRegisterCallback = False
         super().__init__(plm, address, cat, subcat, product_key,
                          description, model)
